requirements.txt
numpy>=1.21.0
torch>=1.9.0
matplotlib>=3.3.0
opencv-python>=4.5.0
pillow>=8.0.0
pyyaml>=5.4.0
swig>=4.0.0